Starting to implement simulated channels, cleaning up scope of some functions

+++ ln_ms_lib/src/sim_channel.rs
@@ -6,6 +6,6 @@ use serde::{Serialize, Deserialize};
 pub struct SimChannel {
     pub src_node: String,
     pub dest_node: String,
-    pub src_balance: i32,
-    pub dest_balance: i32
+    pub src_balance: u64,
+    pub dest_balance: u64
 }
